restructed code

+++ Admin.java
@@ -1,29 +1,57 @@
+package library;
+
+import library.mappers.*;
+import library.entities.*;
+
 import java.sql.*;
-import java.util.ArrayList;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.util.*;
+
+/*
+Description : Admin class operates all the functions/operations of the library,
+such as create user, add books, issue book etc.
+*/
 
 class Admin {
     UserMap userMapper;
     BookMap bookMapper;
+    IssueMap issueMapper;
+    GroupMap groupMapper;
 
+    
     Admin() {
         userMapper = new UserMap();
         bookMapper = new BookMap();
+        issueMapper = new IssueMap();
+        groupMapper = new GroupMap();
     }
 
-    // ----------Admin-User related methods--------
+    // ----------Admin-User related operation methods--------
 
-    // Adding new user
-    public void addUser(String name) {
+    // Adding a new user
+    public void addUser(String name, Integer group) {
         User user = new User();
         user.setName(name);
+        user.setGroup(group);
 
-        User returnUser = userMapper.create(user);
+        User existingUser = userMapper.find(name);
 
-        if (returnUser == null) {
+        if (existingUser != null) {
             System.out.println("User already created by this name.");
+            return ;
+        }
+
+        Integer res = userMapper.create(user);
+
+        if (res == null) {
+            System.out.println("User not added successfully.");
             return;
         }
-        System.out.println("User created. \n " + returnUser);
+
+        User newUser = userMapper.find(name);
+
+        System.out.println("User created. \n " + newUser);
     }
 
     // find user by id
@@ -49,6 +77,12 @@ class Admin {
 
     // remove user by id
     public void removeUser(Integer id) {
+        Integer count = issueMapper.countBooksIssuedByUser(id);
+
+        if ( count != 0) {
+            System.out.println("User has issued some books. Entry can not be deleted.");
+            return;
+        }
         Boolean flag = userMapper.delete(id);
 
         if (flag) {
@@ -59,44 +93,124 @@ class Admin {
 
     }
 
-    // --------Admin-Book related methods---------
+    // --------Admin-Book related operation methods---------
 
-    // Add new book
-    // public void addBook(String name, Integer quantity) {
-    //     Book book = new Book();
-    //     book.setName(name);
-    //     book.setQuantity(quantity);
+    // Add a new book
+    public void addBook(String name, Integer quantity) {
+        Book book = new Book();
+        book.setName(name);
+        book.setLastIssue("1998-03-19");
+        book.setQuantity(quantity);
+        book.setAvailable(quantity);
 
-    //     Book returnBook = bookMapper.create(book);
+        List<Book> existingBook = bookMapper.search(name);
 
-    //     if (returnBook == null) {
-    //         System.out.println("ERROR in Admin.addBook : book not created.");
-    //         return;
-    //     }
-    //     System.out.println("Book created. \n " + returnBook);
-    // }
+        if(!existingBook.isEmpty()) {
+            System.out.println("Book already created by this name.");
+            return ;
+        }
 
-    // public void removeBook(int id) {
-    // db.removeBook(id);
-    // }
+        Integer res = bookMapper.create(book);
 
-    // public void issue(int bookId, int userId) {
-    // db.insert(bookId, userId);
-    // }
+        if(res == 0) {
+            System.out.println("Book not added succesfully");
+            return;
+        }
 
-    // public void returnBook(int bookId, int userId) {
-    // db.removeReturn(bookId, userId);
-    // }
+        List<Book> returnBook = bookMapper.search(name);
+        
+        System.out.println("Book created. \n " + returnBook.get(0));
+    }
 
-    // public void search(String bookName) {
+    // Search similar name matching books
+    public void search(String name) {
+        name = name+"%";
+        
+        List<Book> similarBooks = bookMapper.search(name);
+       
+        if (similarBooks.isEmpty()) {
+            System.out.println("No similar books found.");
+            return;
+        }
 
-    // db.searchSimilar(bookName);
-    // }
+        String label = "";
+        String lastIssueString = "nil";
+        
+        for (Book book : similarBooks) {
+            if ( book.getAvailable() > 0) {
+                label = "Availlable";
+                System.out.println(book + "Status : " + label);
+            } else {
+                label = "Not Available";
+                lastIssueString = book.getLastIssue();
+                System.out.println(book + " Status : " + label + " Will be Available by : " + lastIssueString);
+            }
+        }
 
-    // public void checkUnusedBooks(int days) {
-    // ArrayList<String> bookNames = db.searchUnused(days);
-    // for (String obj : bookNames)
-    // System.out.println(obj);
-    // }
+    }
 
+    //--------Admin - issue/return book related operation methods-------
+
+    //Issue book
+    public void issueBook(Integer userId, Integer bookId) {
+        User existingUser = userMapper.find(userId);
+
+        if (existingUser == null) {
+            System.out.println("No such user");
+            return;
+        }
+
+        Integer groupId = existingUser.getGroup();
+        
+        Group group;
+        Integer count;
+
+        try {
+            group = groupMapper.find(groupId);
+            count = issueMapper.countBooksIssuedByUser(userId);    
+            
+            if ( group == null ) {
+                System.out.println("No such group.");
+                return;
+            }
+        } catch (Exception e) {
+            return;
+        }
+        
+        if (count >  group.getMaxBooks()) {
+            System.out.println("Reached max books limit.");
+            return;
+        }
+
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+        LocalDate issueDate = LocalDate.now();
+        LocalDate dueDate = issueDate.plusDays(group.getMaxTime());
+        String issueDateString = issueDate.format(formatter);
+        String dueDateString = dueDate.format(formatter);
+
+        Issue issue = new Issue();
+        issue.setUserId(userId);
+        issue.setBookId(bookId);
+        issue.setissueDate(issueDateString);
+        issue.setDueDate(dueDateString);
+
+        Issue existingIssue = issueMapper.find(issue);
+
+        if (existingIssue != null ) {
+            System.out.println("User can not issue same book twice.");
+            return;
+        }
+
+        Integer res = issueMapper.create(issue);
+
+        if (res == null) {
+            System.out.println("Book issue failed.");
+            return;
+        }
+
+        Issue newIssue = issueMapper.find(issue);
+
+        Integer result = bookMapper.update(dueDateString, bookId);
+        System.out.println("Issue successful. \n " + newIssue);
+    }
 }
